refactor : #71 - crawler 기능 추상화
